Getting ready for enkf
