+++ Source/Initialization/ERF_InitForEnsemble.cpp
@@ -511,180 +511,6 @@ AddPertToBckgnd(MultiFab& mf_cc_fine,
     }
 }
 
-
-#include <filesystem>
-namespace fs = std::filesystem;
-
-std::string
-get_last_plotfile(const std::string& plotfile_dir)
-{
-    std::vector<std::string> pltfiles;
-
-    for (const auto& entry : fs::directory_iterator(plotfile_dir)) {
-        if (entry.is_directory()) {
-            std::string name = entry.path().filename().string();
-            if (name.find("plt") == 0) {
-                pltfiles.push_back(entry.path().string());
-            }
-        }
-    }
-
-    std::sort(pltfiles.begin(), pltfiles.end());
-    return pltfiles.back();  // last one
-}
-
-
-// Reads the plotfile data into cell cenetred multifab
-// Does not fill ghost cells
-void
-read_plot_file(PlotFileData& pf,
-               const std::vector<std::string> varnames,
-               MultiFab& mf)
-{
-    // ------------------------------------------------------------
-    // Open plotfile
-    // ------------------------------------------------------------
-    const std::vector<std::string>& var_names_pf = pf.varNames();
-
-    // ------------------------------------------------------------
-    // Validate requested variables
-    // ------------------------------------------------------------
-    for (auto const& v : varnames) {
-        bool found = false;
-        for (auto const& vpf : var_names_pf) {
-            if (v == vpf) {
-                found = true;
-                break;
-            }
-        }
-        if (!found) {
-            Abort("read_plot_file: invalid variable name: " + v);
-        }
-    }
-
-    // ------------------------------------------------------------
-    // Define destination MultiFab (single level only)
-    // ------------------------------------------------------------
-    const int level = 0;
-
-    BoxArray ba = pf.boxArray(level);
-    DistributionMapping dm(ba);
-
-    int ncomp = varnames.size();
-
-    mf.define(ba, dm, ncomp, 0);
-
-    // ------------------------------------------------------------
-    // Copy plotfile data → mf
-    // ------------------------------------------------------------
-    for (int comp = 0; comp < ncomp; ++comp)
-    {
-        const MultiFab& src = pf.get(level, varnames[comp]);
-        MultiFab::Copy(mf, src, 0, comp, 1, 0);
-    }
-}
-
-void
-ERF::ComputeAndWriteEnsemblePerturbations()
-{
-    int Nens = solverChoice.n_ensemble;
-    Vector<std::string> varnames = {"density","theta", "x_velocity","y_velocity","z_velocity"};
-    const std::string member_prefix = "member_";
-    // -------------------------------
-    // Step 1: Read last plotfile for each member
-    // -------------------------------
-    Vector<std::unique_ptr<MultiFab>> mf_ens(Nens);
-    std::vector<std::string> pltfiles;
-
-    // Step 1: find all plotfiles (assuming all members have same set)
-    {
-        std::string pf_dir = member_prefix + "00/plotfiles"; // member 0 as reference
-        for (const auto& entry : fs::directory_iterator(pf_dir)) {
-            if (entry.is_directory()) {
-                std::string name = entry.path().filename().string();
-                if (name.find("plt") == 0) {
-                    pltfiles.push_back(name); // store just the directory name
-                }
-            }
-       }
-        std::sort(pltfiles.begin(), pltfiles.end());
-    }
-
-
-// Step 2: loop over all plotfiles (timesteps)
-for (const auto& pf_name : pltfiles)
-{
-    for (int n = 0; n < Nens; ++n)
-    {
-        std::string member_dir = member_prefix + amrex::Concatenate("", n, 2);
-        std::string pf_path = member_dir + "/plotfiles/" + pf_name;
-
-        PlotFileData pf(pf_path);
-
-        const BoxArray& ba = pf.boxArray(0);
-        const DistributionMapping& dm = pf.DistributionMap(0);
-        int ncomp = varnames.size();
-
-        mf_ens[n] = std::make_unique<MultiFab>(ba, dm, ncomp, 0);
-
-        read_plot_file(pf, varnames, *mf_ens[n]);
-    }
-
-    // -------------------------------
-    // Step 2: Compute ensemble mean
-    // -------------------------------
-    MultiFab mf_mean(mf_ens[0]->boxArray(),
-                     mf_ens[0]->DistributionMap(),
-                     mf_ens[0]->nComp(),
-                     mf_ens[0]->nGrow());
-
-    mf_mean.setVal(0.0);
-
-    for (int n = 0; n < Nens; ++n) {
-        MultiFab::Add(mf_mean, *mf_ens[n], 0, 0, mf_mean.nComp(), mf_mean.nGrow());
-    }
-
-    mf_mean.mult(1.0 / Real(Nens), 0, mf_mean.nComp(), mf_mean.nGrow());
-
-    // -------------------------------
-    // Step 3 & 4: Compute perturbations and write plotfiles
-    // -------------------------------
-    for (int n = 0; n < Nens; ++n)
-    {
-        MultiFab mf_pert(mf_mean.boxArray(),
-                         mf_mean.DistributionMap(),
-                         mf_mean.nComp(),
-                         mf_mean.nGrow());
-
-        // perturbation = member - mean
-        MultiFab::Copy(mf_pert, *mf_ens[n], 0, 0, mf_mean.nComp(), mf_mean.nGrow());
-        MultiFab::Subtract(mf_pert, mf_mean, 0, 0, mf_mean.nComp(), mf_mean.nGrow());
-
-        // Create output directory
-        std::string member_dir = member_prefix + amrex::Concatenate("", n, 2);
-        std::string out_dir = member_dir + "/pertfiles";
-        fs::create_directories(out_dir);
-
-        std::string pf_path = member_dir + "/plotfiles/" + pf_name;
-        // Write perturbation plotfile
-        std::string last_pf_name = fs::path(pf_path).filename().string(); // "pltfile00020"
-
-
-        // Extract numeric suffix (everything after "pltfile")
-        std::string suffix = last_pf_name.substr(3);  // "00020"
-
-        // Construct perturbation plotfile name
-        std::string pltname = out_dir + "/plt_pert_" + suffix;
-        WriteSingleLevelPlotfile(pltname,
-                                 mf_pert,
-                                 varnames,
-                                 geom[0],
-                                 0.0,   // time
-                                 0);    // level
-    }
-}
-}
-
 void
 ERF::create_background_state_for_ensemble (int lev,
                                            MultiFab& mf_cc_pert,
